requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
distlib==0.3.5
Django==3.2.12
filelock==3.7.1
idna==3.3
Pillow==9.2.0
pipenv==2022.7.4
platformdirs==2.5.2
psycopg2==2.9.3
pyTelegramBotAPI==4.6.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tornado==6.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
virtualenv==20.15.1
virtualenv-clone==0.5.7
